Move AutotestAutoConfiguration to io.github.arthurc.autotest
We'll have to see if we should create a -core, -web, -junit, etc.
modules but until then we'll just use the "global"
AutotestAutoConfiguration which scans all its submodules for configs.
