+++ src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-io.github.arthurc.autotest.spring.boot.autoconfigure.AutotestAutoConfiguration
+io.github.arthurc.autotest.AutotestAutoConfiguration
