+++ src/test/java/io/github/arthurc/autotest/spring/boot/autoconfigure/ChromeOptionsPropertiesTest.java
@@ -1,5 +1,6 @@
 package io.github.arthurc.autotest.spring.boot.autoconfigure;
 
+import io.github.arthurc.autotest.selenium.ChromeOptionsProperties;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.chrome.ChromeOptions;
